Refactor MIDI/chart writers for maintainability
Addresses the code-quality review of the writer commits (no behavior
change — validated by 435 unit tests + an 8k-chart parse->write->parse
round-trip with zero fidelity drift).

- Decompose the 1458-line midi-writer.ts into a midi-writer/ module:
  index (orchestrator), shared (event model + builders + note helpers),
  events, drums, frets, vocals — each under 500 lines. The public
  `./midi-writer` import path is unchanged (folder index.ts).
- Add typed event builders (metaTextEvent / trackNameEvent / lyricsEvent
  / noteOnEvent / noteOffEvent) in shared.ts, removing ~27 repeated
  `... as MidiEvent` casts at call sites.
- Extract chart/writer-shared.ts (codaTicksFromFreestyle, wrap/unwrap
  event brackets) and use it from both the .chart and .mid writers,
  deleting the duplicated coda-derivation and bracket-normalization logic.
- Replace magic drumType integers (=== 1 / === 2) with the canonical
  drumTypes.fourLanePro / fiveLane constants and type the params DrumType.
- Harden buildUnrecognizedTrack: stable-sort a passthrough track's events
  by absolute tick before delta-encoding, so a malformed (non-monotonic)
  source track no longer crashes the whole write.
